# app-making-book-commercial/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: app-making-book API
  version: "1.0"
  description: |
    Backend API for app-making-book. This spec covers the primary
    endpoints; admin sub-routes not listed here follow the same
    Bearer-auth pattern documented on GET /api/admin/dashboard below —
    see backend/src/routes/adminRoutes.ts for the complete list.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/builder/catalog:
    get:
      summary: List available App Builder modules and templates
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: "{ modules, templates }" } }
  /api/builder/interpret:
    post:
      summary: Natural language -> validated JSON app spec (LLM never outputs code)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties: { prompt: { type: string } }
      responses:
        "200": { description: "{ spec }" }
        "422": { description: The model's output didn't match the required schema }
        "503": { description: GROQ_API_KEY not configured }
  /api/builder/generate:
    post:
      summary: Validated spec -> downloadable zip of real generated code
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [spec]
              properties: { spec: { type: object } }
      responses:
        "200": { description: application/zip file download }
        "400": { description: Spec failed validation }
  /api/builder/presets/{name}:
    get:
      summary: Instant ready-made spec from a preset template (no LLM call)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string, enum: [school, restaurant, crm, ecommerce, inventory] } }
      responses:
        "200": { description: "{ spec }" }
        "404": { description: Unknown preset name }
  /api/builder/regenerate-entity:
    post:
      summary: Regenerate files for ONE entity in an existing spec, not the whole project
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [spec, entityName]
              properties:
                spec: { type: object }
                entityName: { type: string }
                mode: { type: string, enum: [standalone, integrate] }
      responses:
        "200": { description: "{ files }" }
        "422": { description: Entity name not found in the given spec }
  /api/builder/community-templates:
    get:
      summary: Browse public community-shared templates (free sharing, not a paid marketplace)
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Array of public templates } }
    post:
      summary: Share a spec as a community template
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, spec]
              properties:
                name: { type: string }
                description: { type: string }
                spec: { type: object }
                isPublic: { type: boolean }
      responses: { "201": { description: Template saved } }
  /api/builder/community-templates/{id}:
    get:
      summary: Get one community template (increments its download count)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Template with its full spec }
        "404": { description: Not found, or private and not yours }